# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SFP Entity Formation API",
    "description": "Agent-native legal entity formation service. AI agents discover this API via MCP Registry, llms.txt, or OpenAPI spec and invoke it to form Delaware LLCs, Wyoming DAO LLCs, and other entity types on behalf of their human owners. Human involvement is minimized to SSN collection, KYC, and payment approval \u2014 the agent handles everything else.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/entity-orders": {
      "post": {
        "tags": [
          "entity-orders"
        ],
        "summary": "Create entity formation order",
        "description": "Start a new entity formation order. Requires at least one member. Returns the created order with pricing and next required actions.",
        "operationId": "create_order_v1_entity_orders_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "entity-orders"
        ],
        "summary": "List entity orders",
        "description": "Paginated list of entity orders with optional filtering.",
        "operationId": "list_orders_v1_entity_orders_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Per Page"
            },
            "description": "Items per page"
          },
          {
            "name": "jurisdiction",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by jurisdiction code",
              "title": "Jurisdiction"
            },
            "description": "Filter by jurisdiction code"
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by order state",
              "title": "State"
            },
            "description": "Filter by order state"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/entity-orders/{order_id}": {
      "get": {
        "tags": [
          "entity-orders"
        ],
        "summary": "Get entity order",
        "description": "Retrieve a single entity order by ID with next required actions.",
        "operationId": "get_order_v1_entity_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "entity-orders"
        ],
        "summary": "Update entity name",
        "description": "Update the requested entity name. Only allowed in draft or name_check_failed states.",
        "operationId": "update_order_v1_entity_orders__order_id__patch",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/entity-orders/{order_id}/intake": {
      "post": {
        "tags": [
          "entity-orders"
        ],
        "summary": "Complete intake",
        "description": "Mark intake as complete. Transitions draft -> intake_complete. Call this after the order has been created and all member/agent details are finalized.",
        "operationId": "complete_intake_v1_entity_orders__order_id__intake_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateTransitionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/entity-orders/{order_id}/name-check": {
      "post": {
        "tags": [
          "entity-orders"
        ],
        "summary": "Run name availability check",
        "description": "Check whether the requested entity name is available in the target jurisdiction. If available, transitions to name_check_passed. If not, transitions to name_check_failed with suggestions.",
        "operationId": "name_check_v1_entity_orders__order_id__name_check_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NameCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/entity-orders/{order_id}/payment": {
      "post": {
        "tags": [
          "entity-orders"
        ],
        "summary": "Record payment",
        "description": "Record a successful payment. Transitions name_check_passed -> payment_pending, or payment_pending -> payment_complete. In production, this is triggered by the Stripe webhook.",
        "operationId": "record_payment_v1_entity_orders__order_id__payment_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateTransitionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/entity-orders/{order_id}/human-kernel": {
      "post": {
        "tags": [
          "entity-orders"
        ],
        "summary": "Create human kernel session",
        "description": "Create a secure session for the human owner to complete required steps (PII collection, KYC, document signing, attestation). Returns a URL the agent should relay to the human. Transitions order to human_kernel_required.",
        "operationId": "create_kernel_session_v1_entity_orders__order_id__human_kernel_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateKernelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/entity-orders/{order_id}/documents/generate": {
      "post": {
        "tags": [
          "entity-orders"
        ],
        "summary": "Generate formation documents",
        "description": "Generate operating agreement, certificate of formation, and Agent Authority Schedule. Transitions human_kernel_completed -> docs_generated.",
        "operationId": "generate_documents_v1_entity_orders__order_id__documents_generate_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateTransitionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderListResponse": {
        "properties": {
          "orders": {
            "items": {
              "$ref": "#/components/schemas/OrderSummary"
            },
            "type": "array",
            "title": "Orders"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "Total matching orders"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Per Page"
          }
        },
        "type": "object",
        "required": [
          "orders",
          "total",
          "page",
          "per_page"
        ],
        "title": "OrderListResponse",
        "description": "Paginated list of orders."
      },
...
      "UpdateNameRequest": {
        "properties": {
          "requested_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Requested Name",
            "description": "New entity name"
          }
        },
        "type": "object",
        "required": [
          "requested_name"
        ],
        "title": "UpdateNameRequest",
        "description": "Body for PATCH /{order_id} \u2014 update the requested entity name."
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}